NULL terminate char arrays
libvirt uses NULL terminated char arrays in a few places instead of
strings. Interestingly, libvirt does not appear to zero the memory either
before using it.

Return the entire buffer to the caller. This messes up the output of,
e.g., the model in get_node_info, but the caller can get the value
by using hd(binary:split(Model, <<0>>)).