Rework verx_args to be similar to verx_ret
verx_args can convert a proplist of XDR types into a binary that can
be used as argument for the different remote procedure calls. It's
not very useful yet and isn't used right now: it's simpler to look at
the remote_protocol.x file (or run: verx_args:param(<procedure name>)
to get the argument types and their order and construct the binary by
hand. See verx.erl for examples.