Fix tcp transport tests
The tcp transport passes in the PID of the caller. eunit spawns each
test so the spawned test process never receives the messages and times
out (instead, the destroy function would be called in the original
process and get all the messages, leading to some strange errors).

Test pass now.