verx_client_unix: accumulate messages in mailbox
Simulate accumulating messages in the mailbox by sequentially sending
the results of the Unix socket read back to the process.

The reason for doing "passive" reads instead of passing the fd into the
Erlang port poll loop is because of the way the libvirt streams seem
to work. Messages sent to the socket, e.g., after doing a
domain_open_console, never seem to reach libvirtd (the responses are
received). "Passive" reads work for read/writes.

Once I figure out the problem with non-blocking read/writes, the Unix
transport will switch to use erlang:open_port/2.