Update to libvirt 1.2.11
Generate the XDR interface using libvirt git master. Amazingly, the
protocol is backwards compatible and all tests pass on an Ubuntu 12.04
system running libvirt 0.9.8 after some minor code modifications.

The biggest change is that the connection oriented function names are
now namespaced with "connect_". So, where previously the call would be:

    Ref = verx_client:start_link([]),
    verx:open(Ref)

Is now changed to:

    verx:connect_open(Ref)

The documentation, tests and examples have been updated to use the
"connect_" namespace.