mk_libvirt_constants: fail on term conversion
When a value defined in the header file does not have an exact match for
an Erlang term, exit immediately and write the value to stderr.

Remove preceding/trailing whitespace from the name/values. A trailing
tab caused an empty value to be used as the constant.

Note because of the way shell pipes work (unless set -o pipefail is
used), when the output of the script is piped through "sort | uniq", an
error will not cause the whole pipeline to exit, i.e., the output of the
script preceding the error will be written to libvirt_constants.hrl.

Reported by @ksallberg in:

https://github.com/msantos/verx/pull/2

Builds on the fix also provided by ksallberg in:

https://github.com/msantos/verx/pull/3

Thanks Kristian!