rebar3: fix Makefile
Use rebar3 for the makefile targets.

The rebar_alias plugin could be used for the shortcut commands.

Keeping the makefile around for creating the libvirt_constants.hrl and
etags. The libvirt_constants could be moved to the precompile hook with
some changes to ignore if the files do not exist. Building the etags
should be optional.